feature[ats_ZQZZF]: Change identity service to return longs for local ids

+++ plugins/org.eclipse.osee.ats.client.integration.tests/src/org/eclipse/osee/ats/client/integration/tests/ats/core/client/task/createTasks/CreateTasksOperationTest.java
@@ -437,10 +437,10 @@ public class CreateTasksOperationTest {
       }
 
       @Override
-      public int getItemTypeId() {
+      public long getItemTypeId() {
          OseeLog.log(Activator.class, Level.WARNING,
             "CreateTasksOperationTest.MockChange.getItemTypeId() - Unimplemented method called");
-         return 0;
+         return 0L;
       }
 
       @Override
