+++ plugins/org.eclipse.osee.framework.core.model.test/src/org/eclipse/osee/framework/core/model/fields/UniqueIdFieldTest.java
@@ -38,15 +38,13 @@ public class UniqueIdFieldTest extends BaseOseeFieldTest {
    @Parameters
    public static Collection<Object[]> data() {
       Collection<Object[]> data = new ArrayList<Object[]>();
-      data.add(new Object[] {new OseeFieldTestData<Integer>(
-         new UniqueIdField(), //
-         IOseeStorable.UNPERSISTED_VALUE,
-         true, //
-         new FieldGetSetTestData<Integer>(false, IOseeStorable.UNPERSISTED_VALUE, IOseeStorable.UNPERSISTED_VALUE, true), //
-         new FieldGetSetTestData<Integer>(true, IOseeStorable.UNPERSISTED_VALUE, IOseeStorable.UNPERSISTED_VALUE, false), //
-         new FieldGetSetTestData<Integer>(false, 100, 100, true), //
-         new FieldGetSetTestData<Integer>(false, 200, 100, true, OseeStateException.class),//
-         new FieldGetSetTestData<Integer>(true, 300, 100, false, OseeStateException.class)//
+      data.add(new Object[] {new OseeFieldTestData<Long>(new UniqueIdField(), //
+         IOseeStorable.UNPERSISTED_VALUE, true, //
+         new FieldGetSetTestData<Long>(false, IOseeStorable.UNPERSISTED_VALUE, IOseeStorable.UNPERSISTED_VALUE, true), //
+         new FieldGetSetTestData<Long>(true, IOseeStorable.UNPERSISTED_VALUE, IOseeStorable.UNPERSISTED_VALUE, false), //
+         new FieldGetSetTestData<Long>(false, 100L, 100L, true), //
+         new FieldGetSetTestData<Long>(false, 200L, 100L, true, OseeStateException.class),//
+         new FieldGetSetTestData<Long>(true, 300L, 100L, false, OseeStateException.class)//
       )});
       return data;
    }
