+++ plugins/org.eclipse.osee.framework.core.model.test/src/org/eclipse/osee/framework/core/model/mocks/MockOseeTransactionDataAccessor.java
@@ -73,7 +73,7 @@ public class MockOseeTransactionDataAccessor implements ITransactionDataAccessor
    }
 
    @Override
-   public TransactionRecord getOrLoadPriorTransaction(TransactionCache cache, int transactionNumber, int branchId) throws OseeCoreException {
+   public TransactionRecord getOrLoadPriorTransaction(TransactionCache cache, int transactionNumber, long branchId) throws OseeCoreException {
       return null;
    }
