+++ plugins/org.eclipse.osee.framework.core.model/src/org/eclipse/osee/framework/core/model/BranchFactory.java
@@ -54,7 +54,7 @@ public class BranchFactory implements IOseeTypeFactory {
       return branch;
    }
 
-   public Branch createOrUpdate(IOseeCache<String, Branch> cache, int uniqueId, StorageState storageState, String guid, String name, BranchType branchType, BranchState branchState, boolean isArchived) throws OseeCoreException {
+   public Branch createOrUpdate(IOseeCache<String, Branch> cache, long uniqueId, StorageState storageState, String guid, String name, BranchType branchType, BranchState branchState, boolean isArchived) throws OseeCoreException {
       Conditions.checkNotNull(cache, "BranchCache");
       Branch branch = cache.getById(uniqueId);
       if (branch == null) {
